requirements.txt
django 
networkx 
spacy 
lexrank 
distance 
pattern 
autocorrect 
inflect 
contractions 
hdbscan 
nltk
scipy
scikit-learn
powerai
sentencepiece
numpy
tqdm
transformers==2.3.0
sentence_transformers 
python-dateutil 
pillow